move yaml handler to separate file
