+++ tests/test.py
@@ -2,6 +2,7 @@ import unittest
 import os
 
 from bedshift import bedshift
+from bedshift import BedshiftYAMLHandler
 
 
 class TestBedshift(unittest.TestCase):
@@ -68,27 +69,6 @@ class TestBedshift(unittest.TestCase):
         self.assertEqual(dropped, 0)
         self.bs.reset_bed()
 
-    def test_handle_yaml(self):
-        yamled = self.bs.handle_yaml(bedshifter=self.bs, yaml_fp="tests/bedshift_analysis.yaml")
-        self.bs.reset_bed()
-
-        added = self.bs.add(addrate=0.1, addmean=100, addstdev=20)
-        f_drop_10 = self.bs.drop_from_file(fp="tests/test.bed", droprate=0.1)
-        f_shift_30 = self.bs.shift_from_file(fp="tests/bedshifted_test.bed",
-                                            shiftrate=0.50,
-                                            shiftmean=100,
-                                            shiftstdev=200)
-        f_added_20 = self.bs.add_from_file(fp="tests/small_test.bed", addrate=0.2)
-        cut = self.bs.cut(cutrate=0.2)
-        shifted = self.bs.shift(shiftrate=0.3, shiftmean=100, shiftstdev=200)
-        dropped = self.bs.drop(droprate=0.3)
-        merged = self.bs.merge(mergerate=0.15)
-
-        total = added+f_drop_10+f_shift_30+f_added_20+cut+dropped+shifted+merged
-
-        self.assertAlmostEqual(yamled, total, places=-2)
-        self.bs.reset_bed()
-
     def test_all_perturbations1(self):
         perturbed = self.bs.all_perturbations(
                             addrate=0.5, addmean=320.0, addstdev=20.0,
@@ -127,3 +107,25 @@ class TestBedshift(unittest.TestCase):
         added = bs_small.add(2.0, 100, 50)
         self.assertEqual(added, 4)
 
+
+class TestBedshiftYAMLHandler(unittest.TestCase):
+    def test_handle_yaml(self):
+        yamled = self.bs.handle_yaml(bedshifter=self.bs, yaml_fp="tests/bedshift_analysis.yaml")
+        self.bs.reset_bed()
+
+        added = self.bs.add(addrate=0.1, addmean=100, addstdev=20)
+        f_drop_10 = self.bs.drop_from_file(fp="tests/test.bed", droprate=0.1)
+        f_shift_30 = self.bs.shift_from_file(fp="tests/bedshifted_test.bed",
+                                            shiftrate=0.50,
+                                            shiftmean=100,
+                                            shiftstdev=200)
+        f_added_20 = self.bs.add_from_file(fp="tests/small_test.bed", addrate=0.2)
+        cut = self.bs.cut(cutrate=0.2)
+        shifted = self.bs.shift(shiftrate=0.3, shiftmean=100, shiftstdev=200)
+        dropped = self.bs.drop(droprate=0.3)
+        merged = self.bs.merge(mergerate=0.15)
+
+        total = added+f_drop_10+f_shift_30+f_added_20+cut+dropped+shifted+merged
+
+        self.assertAlmostEqual(yamled, total, places=-2)
+        self.bs.reset_bed()
